requirements.txt
# Version: 3.1.0
# Base requirements for DocuMind-HR microservice
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pypdf
pydantic
sentence-transformers==2.7.0
numpy>=1.24
